emotion-recognition/requirements.txt
# ============================================================
# Core ML / Deep Learning
# ============================================================
# TensorFlow 2.15 CPU build – compatible with Python 3.10 on Windows
tensorflow==2.15.0
tensorflow-io-gcs-filesystem==0.31.0

# ============================================================
# Numeric + Data + Evaluation
# ============================================================
numpy>=1.26.0,<2.0.0
pandas>=2.0.0,<2.3.0
scikit-learn>=1.3,<1.6
matplotlib>=3.7,<3.9
Pillow>=10.0.0
tqdm>=4.66.0

# ============================================================
# App / Demo
# ============================================================
streamlit>=1.32.0,<2.0.0

# Optional face detection
mtcnn==0.1.1

